Attempting Stack data structure with F#
First, I've attempted to analyze the logic of the code and understand how it suppose to work. I've compiled and ran the code to see if any compilation errors come up, and none do; however, I can't get it to output anything yet to if see understand the true inner workings of the code.